@@ -6,7 +6,7 @@ TEST_CASE("TimSortSimpleTest")
 {
     int arr[] = { 3, 2, 1 };
 
-    leviathan::algorithm::ranges::tim_sort(arr);
+    leviathan::algorithm::tim_sort(arr);
 
     REQUIRE(arr[0] == 1);
     REQUIRE(arr[1] == 2);
